backend/pyproject.toml
[tool.poetry]
name = "poca"
version = "0.1.0"
description = ""
authors = ["MU-Software <musoftware@daum.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pydantic = "^2.8.2"
pillow = "^10.4.0"
aiohttp = "^3.10.5"
playwright = "^1.46.0"
opencv-python = "^4.10.0.84"
websockets = "^12.0"
jinja2 = "^3.1.4"
pyserial = "^3.5"
redis = "^5.0.8"
pydantic-settings = "^2.4.0"
typer = "^0.12.5"
async-lru = "^2.0.4"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pylint = "^3.2.7"
flake8 = "^7.1.1"
mypy = "^1.11.2"
black = "^24.8.0"
bandit = "^1.7.9"
isort = "^5.13.2"
refurb = "^2.0.0"
flake8-noqa = "^1.4.0"
flake8-bugbear = "^24.8.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
